requirements.txt
pandas==1.3.1
torchvision==0.10.0
torch==1.9.0
numpy==1.21.1
requests==2.26.0
Pillow==8.3.1
scikit_learn==0.24.2
